Use string representation of double in order not to lose location data

+++ domain/src/main/kotlin/com/ringoid/domain/manager/ISharedPrefsManager.kt
@@ -41,7 +41,7 @@ interface ISharedPrefsManager {
 
     /* Location */
     // ------------------------------------------
-    fun getLocation(): Pair<Float, Float>?
+    fun getLocation(): Pair<Double, Double>?
 
     fun saveLocation(location: Location)
